don't use floats

+++ src/main/java/org/spongepowered/api/item/recipe/furnace/FurnaceRecipe.java
@@ -29,6 +29,6 @@ import org.spongepowered.api.item.recipe.Recipe;
 
 public interface FurnaceRecipe extends Recipe {
 
-    float getExperience(GridInventory grid);
+    double getExperience(GridInventory grid);
 
 }
